Remake of the CALIFA simulation
The CALIFA simulation is here rebuilt based on CalifaPoints and a new R3BCalifaDigitizer.
A few errors from the previous scheme has been corrected and old code has been removed.
The complete (and final) BARREL+iPhos geometry is now included (2020) and all other geometries
are removed (except demonstrator geometries based on parts of 2020 geometry).
The main modifications are listed below:
 - R3BCalifa is now simpler and fills CalifaPoints, similar to other detectors.
 - CalifaCrystalCalDataSim and CalifaHitDataSim are removed. kSimulation flag has been removed
in several classes as the data treatment is homogeneous (dealing with sim and real data in the
same way from CalifaCrystalCalData).
 - Errors related with the use of iNode volume identifiers are solved by using the crystal path.
Unnecessary structs or std::maps have been removed.
 - A new digitizer (R3BCalifaDigitizer) fills CalifaCrystalCalData from CalifaPoints. Now the
crystal non-uniformity and resolution are defined in the digitizer and removed from the Cal2Hit.
 - A new geometry ("2020") contains the BARREL and iPhos alveoli and crystals (including
wrapping) in their final (hopefully) version. A simpler macro and name scheme to identify
crystals and all other volumes are included.
 - Old geometries and perl macros to produce them are removed.
 - Consequently, R3BCalifaGeometry is much simpler and clear, with a set of functions which are
used in different reconstruction levels to obtain the crystalId of the volumes or to get the
position parameter (angles) for each crystal.